+++ src/main/java/com/yanny/ytech/ModBusSubscriber.java
@@ -3,7 +3,6 @@ package com.yanny.ytech;
 import com.yanny.ytech.compatibility.CuriosCapability;
 import com.yanny.ytech.compatibility.TopCompatibility;
 import com.yanny.ytech.configuration.SpearType;
-import com.yanny.ytech.configuration.block_entity.IrrigationBlockEntity;
 import com.yanny.ytech.configuration.block_entity.WellPulleyBlockEntity;
 import com.yanny.ytech.configuration.entity.*;
 import com.yanny.ytech.configuration.item.BasketItem;
@@ -16,7 +15,10 @@ import com.yanny.ytech.configuration.screen.PrimitiveAlloySmelterScreen;
 import com.yanny.ytech.configuration.screen.PrimitiveSmelterScreen;
 import com.yanny.ytech.network.irrigation.IrrigationServerNetwork;
 import com.yanny.ytech.network.irrigation.IrrigationUtils;
-import com.yanny.ytech.registration.*;
+import com.yanny.ytech.registration.YTechBlockEntityTypes;
+import com.yanny.ytech.registration.YTechEntityTypes;
+import com.yanny.ytech.registration.YTechItems;
+import com.yanny.ytech.registration.YTechMenuTypes;
 import net.minecraft.client.gui.screens.MenuScreens;
 import net.minecraft.client.renderer.entity.ThrownItemRenderer;
 import net.minecraft.client.renderer.item.ItemProperties;
@@ -193,17 +195,15 @@ public class ModBusSubscriber {
 
     @SubscribeEvent
     public static void onRegisterCap(@NotNull RegisterCapabilitiesEvent event) {
-        event.registerBlock(Capabilities.FluidHandler.BLOCK, (level, pos, state, be, side) -> {
-            if (!level.isClientSide && be instanceof IrrigationBlockEntity irrigationBlockEntity) {
-                IrrigationServerNetwork network = YTechMod.IRRIGATION_PROPAGATOR.server().getNetwork(irrigationBlockEntity);
+        event.registerBlockEntity(Capabilities.FluidHandler.BLOCK, YTechBlockEntityTypes.AQUEDUCT.get(), (blockEntity, direction) -> {
+            IrrigationServerNetwork network = YTechMod.IRRIGATION_PROPAGATOR.server().getNetwork(blockEntity);
 
-                if (network != null) {
-                    return network.getFluidHandler();
-                }
+            if (network != null) {
+                return network.getFluidHandler();
             }
 
             return null;
-        }, YTechBlocks.AQUEDUCT.get());
+        });
         event.registerBlockEntity(Capabilities.ItemHandler.BLOCK, YTechBlockEntityTypes.AMPHORA.get(), (amphoraBlockEntity, direction) -> {
             if (!amphoraBlockEntity.isRemoved() && direction == Direction.UP) {
                 return amphoraBlockEntity.getItemHandler();