+++ src/main/java/com/yanny/ytech/configuration/block/AqueductBlock.java
@@ -1,12 +1,12 @@
 package com.yanny.ytech.configuration.block;
 
 import com.yanny.ytech.YTechMod;
+import com.yanny.ytech.configuration.MaterialType;
 import com.yanny.ytech.configuration.Utils;
 import com.yanny.ytech.configuration.block_entity.AqueductBlockEntity;
 import com.yanny.ytech.configuration.recipe.RemainingShapedRecipe;
 import com.yanny.ytech.network.irrigation.IrrigationClientNetwork;
 import com.yanny.ytech.network.irrigation.IrrigationServerNetwork;
-import com.yanny.ytech.registration.YTechBlocks;
 import com.yanny.ytech.registration.YTechItemTags;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
@@ -21,6 +21,7 @@ import net.minecraft.util.RandomSource;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.Items;
 import net.minecraft.world.item.context.BlockPlaceContext;
@@ -42,11 +43,14 @@ import net.minecraft.world.phys.shapes.CollisionContext;
 import net.minecraft.world.phys.shapes.Shapes;
 import net.minecraft.world.phys.shapes.VoxelShape;
 import net.neoforged.neoforge.client.model.generators.BlockStateProvider;
+import net.neoforged.neoforge.client.model.generators.ModelBuilder;
 import net.neoforged.neoforge.client.model.generators.ModelFile;
 import net.neoforged.neoforge.client.model.generators.MultiPartBlockStateBuilder;
 import net.neoforged.neoforge.fluids.FluidStack;
 import net.neoforged.neoforge.fluids.capability.IFluidHandler;
 import net.neoforged.neoforge.fluids.capability.templates.FluidTank;
+import net.neoforged.neoforge.registries.DeferredBlock;
+import net.neoforged.neoforge.registries.DeferredItem;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
@@ -74,13 +78,23 @@ public class AqueductBlock extends IrrigationBlock implements BucketPickup, Liqu
     private static final BooleanProperty SOUTH_WEST = BooleanProperty.create("south_west");
 
     private final Map<BlockState, VoxelShape> shapesCache;
+    private final MaterialType material;
 
-    public AqueductBlock() {
+    public AqueductBlock(MaterialType material) {
         super(Properties.ofFullCopy(Blocks.TERRACOTTA));
         this.shapesCache = this.getShapeForEachState(AqueductBlock::calculateShape);
+        this.material = material;
     }
 
-    @SuppressWarnings("deprecation")
+    public int getCapacity() {
+        return switch (material) {
+            case MUDBRICK -> YTechMod.CONFIGURATION.getBaseFluidStoragePerBlock();
+            case TERRACOTTA -> YTechMod.CONFIGURATION.getBaseFluidStoragePerBlock() * 2;
+            case STONE -> YTechMod.CONFIGURATION.getBaseFluidStoragePerBlock() * 4;
+            default -> throw new IllegalStateException(String.format("Missing material type %s", material));
+        };
+    }
+
     @NotNull
     @Override
     public RenderShape getRenderShape(@NotNull BlockState blockState) {
@@ -149,7 +163,7 @@ public class AqueductBlock extends IrrigationBlock implements BucketPickup, Liqu
     @NotNull
     @Override
     public BlockEntity newBlockEntity(@NotNull BlockPos pos, @NotNull BlockState blockState) {
-        return new AqueductBlockEntity(pos, blockState);
+        return new AqueductBlockEntity(pos, blockState, getCapacity());
     }
 
     @NotNull
@@ -161,7 +175,7 @@ public class AqueductBlock extends IrrigationBlock implements BucketPickup, Liqu
             if (network != null) {
                 FluidTank tank = network.getFluidHandler();
 
-                if (tank.getFluidAmount() > 1000) {
+                if (tank.getFluidAmount() >= 1000) {
                     tank.drain(1000, IFluidHandler.FluidAction.EXECUTE);
                     return new ItemStack(Items.WATER_BUCKET);
                 }
@@ -260,9 +274,17 @@ public class AqueductBlock extends IrrigationBlock implements BucketPickup, Liqu
         return Direction.Plane.HORIZONTAL.stream().map((dir) -> pos.offset(dir.getNormal())).toList();
     }
 
-    public static void registerModel(@NotNull BlockStateProvider provider) {
-        ResourceLocation bricks = Utils.modBlockLoc("terracotta_bricks");
-        String name = Utils.getPath(YTechBlocks.AQUEDUCT);
+    public static void registerModel(@NotNull BlockStateProvider provider, @NotNull DeferredBlock<Block> block, MaterialType material) {
+        ResourceLocation bricks;
+
+        switch (material) {
+            case MUDBRICK -> bricks = Utils.mcBlockLoc("mud_bricks");
+            case TERRACOTTA -> bricks = Utils.modBlockLoc("terracotta_bricks");
+            case STONE -> bricks = Utils.mcBlockLoc("stone_bricks");
+            default -> throw new IllegalStateException();
+        }
+
+        String name = Utils.getPath(block);
         ModelFile base = provider.models().getBuilder(name)
                 .parent(provider.models().getExistingFile(Utils.mcBlockLoc("block")))
                 .element().allFaces((direction, faceBuilder) -> {
@@ -281,7 +303,7 @@ public class AqueductBlock extends IrrigationBlock implements BucketPickup, Liqu
                         case NORTH, SOUTH -> faceBuilder.uvs(2, 0, 14, 14).texture("#0");
                         case EAST -> faceBuilder.uvs(14, 0, 16, 14).texture("#0");
                         case WEST -> faceBuilder.uvs(0, 0, 2, 14).texture("#0");
-                        case UP -> faceBuilder.uvs(2, 0, 14, 2).texture("#0");
+                        case UP -> faceBuilder.uvs(2, 0, 14, 2).rotation(ModelBuilder.FaceRotation.UPSIDE_DOWN).texture("#0");
                     }
                 })
                 .from(2, 2, 0).to(14, 16, 2).end()
@@ -293,14 +315,14 @@ public class AqueductBlock extends IrrigationBlock implements BucketPickup, Liqu
                     switch(direction) {
                         case NORTH, EAST -> faceBuilder.uvs(14, 0, 16, 14).texture("#0");
                         case SOUTH, WEST -> faceBuilder.uvs(0, 0, 2, 14).texture("#0");
-                        case UP -> faceBuilder.uvs(0, 0, 2, 2).texture("#0");
+                        case UP -> faceBuilder.uvs(2, 0, 4, 2).texture("#0");
                     }
                 })
                 .from(0, 2, 0).to(2, 16, 2).end()
                 .texture("particle", bricks)
                 .texture("0", bricks);
 
-        MultiPartBlockStateBuilder builder = provider.getMultipartBuilder(YTechBlocks.AQUEDUCT.get()).part().modelFile(base).addModel().end();
+        MultiPartBlockStateBuilder builder = provider.getMultipartBuilder(block.get()).part().modelFile(base).addModel().end();
 
         PROPERTY_BY_DIRECTION.forEach((dir, value) -> builder.part().modelFile(side).rotationY(ANGLE_BY_DIRECTION.get(dir)).addModel().condition(value, true).end());
         builder.part().modelFile(edge).rotationY(ANGLE_BY_DIRECTION.get(Direction.NORTH)).addModel().condition(NORTH_WEST, true).end();
@@ -341,14 +363,33 @@ public class AqueductBlock extends IrrigationBlock implements BucketPickup, Liqu
         provider.itemModels().getBuilder(name).parent(itemModel);
     }
 
-    public static void registerRecipe(RecipeOutput recipeConsumer) {
-        RemainingShapedRecipe.Builder.shaped(RecipeCategory.MISC, YTechBlocks.AQUEDUCT.get())
-                .define('#', YTechItemTags.TERRACOTTA_BRICKS)
-                .pattern("# #")
-                .pattern("# #")
-                .pattern("###")
-                .unlockedBy(Utils.getHasName(), RecipeProvider.has(YTechItemTags.TERRACOTTA_BRICKS))
-                .save(recipeConsumer, Utils.modLoc(YTechBlocks.AQUEDUCT));
+    public static void registerRecipe(RecipeOutput recipeConsumer, DeferredItem<Item> item, MaterialType material) {
+        switch (material) {
+            case MUDBRICK -> RemainingShapedRecipe.Builder.shaped(RecipeCategory.MISC, item.get(), 2)
+                    .define('#', Items.MUD_BRICKS)
+                    .pattern("# #")
+                    .pattern("# #")
+                    .pattern("###")
+                    .unlockedBy(Utils.getHasName(), RecipeProvider.has(Items.MUD_BRICKS))
+                    .save(recipeConsumer, Utils.modLoc(item));
+            case TERRACOTTA -> RemainingShapedRecipe.Builder.shaped(RecipeCategory.MISC, item.get(), 2)
+                    .define('#', YTechItemTags.TERRACOTTA_BRICKS)
+                    .define('H', YTechItemTags.HAMMERS.tag)
+                    .pattern("# #")
+                    .pattern("#H#")
+                    .pattern("###")
+                    .unlockedBy(Utils.getHasName(), RecipeProvider.has(YTechItemTags.TERRACOTTA_BRICKS))
+                    .save(recipeConsumer, Utils.modLoc(item));
+            case STONE -> RemainingShapedRecipe.Builder.shaped(RecipeCategory.MISC, item.get(), 2)
+                    .define('#', Items.STONE_BRICKS)
+                    .define('H', YTechItemTags.HAMMERS.tag)
+                    .pattern("# #")
+                    .pattern("#H#")
+                    .pattern("###")
+                    .unlockedBy(Utils.getHasName(), RecipeProvider.has(Items.STONE_BRICKS))
+                    .save(recipeConsumer, Utils.modLoc(item));
+            default -> throw new IllegalStateException();
+        }
     }
 
     private static boolean hasSide(@NotNull LevelAccessor level, @NotNull BlockPos pos, @NotNull BooleanProperty property) {