+++ src/main/java/com/yanny/ytech/configuration/block_entity/AqueductBlockEntity.java
@@ -6,20 +6,28 @@ import com.yanny.ytech.network.irrigation.NetworkType;
 import com.yanny.ytech.registration.YTechBlockEntityTypes;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Holder;
+import net.minecraft.nbt.CompoundTag;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.level.biome.Biome;
 import net.minecraft.world.level.block.state.BlockState;
 import org.jetbrains.annotations.NotNull;
 
 public class AqueductBlockEntity extends IrrigationBlockEntity {
+    public static final String TAG_CAPACITY = "capacity";
+
+    private int capacity;
 
     public AqueductBlockEntity(@NotNull BlockPos pos, @NotNull BlockState blockState) {
-        super(YTechBlockEntityTypes.AQUEDUCT.get(), pos, blockState, ((AqueductBlock)blockState.getBlock()).getValidNeighbors(blockState, pos));
+        this(pos, blockState, 0);
     }
 
-    @Override
-    public int getFlow() {
-        return 0;
+    public AqueductBlockEntity(@NotNull BlockPos pos, @NotNull BlockState blockState, int capacity) {
+        super(YTechBlockEntityTypes.AQUEDUCT.get(), pos, blockState, ((AqueductBlock)blockState.getBlock()).getValidNeighbors(blockState, pos));
+        this.capacity = capacity;
+    }
+
+    public int getCapacity() {
+        return capacity;
     }
 
     @Override
@@ -38,7 +46,19 @@ public class AqueductBlockEntity extends IrrigationBlockEntity {
         return false;
     }
 
+    @Override
+    public void load(@NotNull CompoundTag tag) {
+        super.load(tag);
+        capacity = tag.getInt(TAG_CAPACITY);
+    }
+
     public void onRandomTick() {
         YTechMod.IRRIGATION_PROPAGATOR.server().changed(this);
+    }
+
+    @Override
+    protected void saveAdditional(@NotNull CompoundTag tag) {
+        super.saveAdditional(tag);
+        tag.putInt(TAG_CAPACITY, capacity);
     }
 }