+++ src/main/java/com/yanny/ytech/configuration/block_entity/AqueductValveBlockEntity.java
@@ -55,11 +55,6 @@ public class AqueductValveBlockEntity extends IrrigationBlockEntity {
     }
 
     @Override
-    public boolean validForRainFilling() {
-        return false;
-    }
-
-    @Override
     public @NotNull NetworkType getNetworkType() {
         return NetworkType.PROVIDER;
     }