+++ src/main/java/com/yanny/ytech/configuration/block_entity/WellPulleyBlockEntity.java
@@ -14,6 +14,7 @@ import net.minecraft.nbt.CompoundTag;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.sounds.SoundSource;
 import net.minecraft.world.InteractionResult;
+import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.biome.Biome;
 import net.minecraft.world.level.block.Block;
@@ -52,11 +53,6 @@ public class WellPulleyBlockEntity extends IrrigationBlockEntity {
         return 0;
     }
 
-    @Override
-    public boolean validForRainFilling() {
-        return false;
-    }
-
     @NotNull
     @Override
     public NetworkType getNetworkType() {
@@ -64,7 +60,7 @@ public class WellPulleyBlockEntity extends IrrigationBlockEntity {
     }
 
     @NotNull
-    public InteractionResult onUse(@NotNull Level level, @NotNull BlockState aboveState) {
+    public InteractionResult onUse(@NotNull Level level, @NotNull BlockState aboveState, @NotNull Player player) {
         if (aboveState.is(YTechBlocks.WELL_PULLEY.get())) {
             timer = 20;
 
@@ -75,6 +71,7 @@ public class WellPulleyBlockEntity extends IrrigationBlockEntity {
 
             level.playSound(null, worldPosition, YTechSoundEvents.WELL_PULLEY_USE.get(), SoundSource.BLOCKS, 1.0f, 0.8f + level.random.nextFloat() * 0.4f);
             level.blockEntityChanged(worldPosition);
+            player.causeFoodExhaustion(0.5f);
             return InteractionResult.CONSUME;
         }