+++ src/main/java/com/yanny/ytech/generation/YTechLootTableProvider.java
@@ -63,7 +63,6 @@ class YTechLootTableProvider extends LootTableProvider {
         @Override
         protected void generate() {
             dropSelf(YTechBlocks.AMPHORA);
-            dropSelf(YTechBlocks.AQUEDUCT);
             dropSelf(YTechBlocks.AQUEDUCT_FERTILIZER);
             dropSelf(YTechBlocks.AQUEDUCT_HYDRATOR);
             dropSelf(YTechBlocks.AQUEDUCT_VALVE);
@@ -88,6 +87,7 @@ class YTechLootTableProvider extends LootTableProvider {
             registerWellPulleyLootTable();
             dropSelf(YTechBlocks.WOODEN_BOX);
 
+            registerMaterialLootTable(YTechBlocks.AQUEDUCTS, this::dropSelf);
             registerMaterialLootTable(YTechBlocks.DEEPSLATE_ORES, this::oreLoot, MaterialType.VANILLA_METALS);
             registerMaterialLootTable(YTechBlocks.DRYING_RACKS, this::dropSelf);
             registerMaterialLootTable(YTechBlocks.GRAVEL_DEPOSITS, (block, material) -> depositLootProvider(block, material, Items.GRAVEL));
@@ -105,7 +105,6 @@ class YTechLootTableProvider extends LootTableProvider {
             return Stream.of(
                     Stream.of(
                             YTechBlocks.AMPHORA,
-                            YTechBlocks.AQUEDUCT,
                             YTechBlocks.AQUEDUCT_FERTILIZER,
                             YTechBlocks.AQUEDUCT_HYDRATOR,
                             YTechBlocks.AQUEDUCT_VALVE,
@@ -130,6 +129,7 @@ class YTechLootTableProvider extends LootTableProvider {
                             YTechBlocks.WELL_PULLEY,
                             YTechBlocks.WOODEN_BOX
                     ).map(DeferredBlock::get),
+                    YTechBlocks.AQUEDUCTS.entries().stream().map(Map.Entry::getValue).map(DeferredBlock::get),
                     filteredStream(YTechBlocks.DEEPSLATE_ORES, MaterialType.VANILLA_METALS).map(Map.Entry::getValue).map(DeferredBlock::get),
                     YTechBlocks.DRYING_RACKS.entries().stream().map(Map.Entry::getValue).map(DeferredBlock::get),
                     YTechBlocks.GRAVEL_DEPOSITS.entries().stream().map(Map.Entry::getValue).map(DeferredBlock::get),